extend functionality to real numbers

+++ main.py
@@ -5,6 +5,6 @@ if __name__ == "__main__":
         ops = command.split(' ')
         # recognize the command
         if ops[0] == "+":
-            print(int(ops[1]) + int(ops[2]))
+            print(float(ops[1]) + float(ops[2]))
         command = input().strip()
     print("Thanks for choosing us :)")
